Xlint:all the things (and -Werror)
Enable lint warnings in javac and treat them as errors. Also fix the
already existing warnings.

+++ src/main/java/de/measite/minidns/DNSMessage.java
@@ -51,7 +51,7 @@ public class DNSMessage {
          * @return the response code.
          */
         public byte getValue() {
-            return (byte) value;
+            return value;
         }
 
         /**
